Add support for nested prefix matching on anything but (#1)
The compiler did not have a match case for the
rule_nested_prefix_matching
grammar of the event. Let it fall through to same as prefix.

Added a test for nested prefix under anything-but condition and
found a bug in the compiler. Added a negation to the output rule
to return when the rule does not match.