feat: allow compile to accept a single file or folder
Compile can handle a single file or a folder containing .json
event rules. Compile always returns an array of rego bodies.

The rego policy name is unchanged if given a single file. However,
if a folder is provided each policy is given the package name of
rule2rego.<relativePathAndName>. An additional policy is generated
that encapsulates all policies to check if _any_ policy matches
the input. This package is given the name rule2rego and works the
exact same way as a single policy file.