feat: folder input support (#6)
* feat: allow compile to accept a single file or folder

Compile can handle a single file or a folder containing .json
event rules. Compile always returns an array of rego bodies.

The rego policy name is unchanged if given a single file. However,
if a folder is provided each policy is given the package name of
rule2rego.<relativePathAndName>. An additional policy is generated
that encapsulates all policies to check if _any_ policy matches
the input. This package is given the name rule2rego and works the
exact same way as a single policy file.

* chore: Add a contributing guide

Copied information from the readme about adding tests to the
contributing guide. Also added a little information about setting
up the repository.

* fix: add automatically removed trailing whitespace in readme

VSCode removed the trailing whitespace. Set the local settings
so that whitespace is not trimmed.

Removed extra whitespace in the README.

* fix: use destructuring in global setup for legibility

[policyDocument] reads better than policyDocuments and later
access by index at policyDocuments[0] when we only ever have a single
policy document from the preceding glob + map.