Add support for nested arrays in input (#8)
This issue comes up when using nested arrays which may be
possible using custom events. Unaware of any cases where the
AWS API internally will send an input payload with nested arrays
but the cloudwatch event rule patterns do allow for it when matching
value. The simplest fix was to treat all input as if each key
is an array by checking and converting them to arrays.