fix: exist fails on lists and booleans
The old check looked if the key was set AND truthy. Meaning "key: false"
would deny when it should allow. When I introduced the array
colescence in the last PR I further broke the exist matcher
by producing rego that failed on certain input formats but otherwise
looked to perform correctly in these tests.

New change coalesces the value to an array, iterates the array,
and checks if any index contains the key. A new helper function
is added to these files as needed to check for the existence
of the key rather than checking the key or value itself.